app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pytest
pytest-mock

